Labo/01_labo-user-initial/src/main/java/be/ucll/backend/users/domain/User.java
package be.ucll.backend.users.domain;

import java.util.ArrayList;
import java.util.List;

public class User {

    private String name;
    private int age;
    private List<Integer> membershipYears = new ArrayList<Integer>();

    public User(String name, int age) {
        this.name = name;
        if (age >= 0)
            this.age = age;
    }

    public User() {
    }

    public int getAge() {
        return this.age;
    }

    public void setAge(int age) {
        if (age > 0) {
            this.age = age;
        }
    }

    public String getName() {
        return name;
    }

    public int countYearsOfMembership() {
        return membershipYears.size();
    }

    public void addMembershipYear(int year) {
        membershipYears.add(year);
    }

    public int countMembershipYearsAfter1999() {
        int result = 0;
        for (Integer year : membershipYears) {
            if (year > 1999)
                result++;
        }
        return result;
    }

    public boolean equals(User otherUser) {
        return this.name.equals(otherUser.getName()) && this.age == otherUser.getAge();
    }

}